+++ Units/parser-cmake.r/cmake-function.d/input.cmake
@@ -1,0 +1,19 @@
+function(tag_this1)
+endfunction(tag_this1)
+	function(	Bfunctiontag_this_2 ${BAR})
+
+function(cxx_shared_library name cxx_flags)
+  cxx_library_with_type(${name} SHARED "${cxx_flags}" ${ARGN})
+endfunction()
+
+    FUNCTION		(
+        d_ALSO_Tag_this
+    true)
+
+Function(
+eLastlyTagThis_ "hello")
+
+function(${not_this} foo)
+function(not-this foo)
+function(not.this foo)
+function(1notthis foo)