+++ Units/parser-cmake.r/cmake-option.d/input.cmake
@@ -1,0 +1,15 @@
+option(1_tag_this foo)
+option(B-option-tag_this_2 ${BAR})
+
+    option ( c_TAG_THIS but_not_this)
+
+    OPTION		(
+        d-ALSO-Tag.this
+    true)
+
+Option(
+e.LastlyTagThis "hello")
+
+option(${not_this} foo)
+option(${ not_this } foo)
+option(not~this foo)