+++ Units/parser-cmake.r/cmake-project.d/input.cmake
@@ -1,0 +1,15 @@
+project(1_tag_this)
+project(B-project-tag_this_2 ${BAR})
+
+    project ( c_TAG_THIS but_not_this)
+
+    PROJECT		(
+        d-ALSO-Tag.this
+	)
+
+Project(
+e.LastlyTagThis "hello")
+
+project(${not_this})
+project(${ not_this })
+project(not/this)