+++ Units/parser-cmake.r/cmake-variable.d/input.cmake
@@ -1,0 +1,13 @@
+set(1_tag_this foo)
+set(B-set-tag_this_2 ${BAR})
+
+    set ( c_TAG_THIS but_not_this)
+
+    SET		(
+        d-ALSO-Tag.this
+    true)
+
+Set(
+e.LastlyTagThis "hello")
+
+set(${not_this} foo)