+++ main/parsers.h
@@ -47,6 +47,7 @@
 	BasicParser, \
 	BetaParser, \
 	ClojureParser, \
+	CMakeParser, \
 	CParser, \
 	CppParser, \
 	CPreProParser, \